// src/models/interfaces/Item.ts
import { LocalizedString } from './LocalizedString';
import { Venue } from './Venue';
import { ItemType } from './ItemType';

export interface Item {
    type: ItemType;
    id: string;
    deleted: boolean;
    chat_id?: string;
    chronological_id?: string;
    created_at: Date;
    created_by?: string;
    edited_at?: Date;
    edited_by?: string;
    external_id?: string;
    externally_deleted?: boolean;
    feed_text?: LocalizedString;
    feed_title?: LocalizedString;
    short_description?: LocalizedString;
    status?: string;
    title?: LocalizedString;
    video_url?: string;
    deal_type?: string;
    domains?: string[];
    starts_at?: Date;
    gallery?: string[];
    no_of_tickets_left?: number;
    venue?: Venue;
    author_id?: string;
    organization_id?: string;
    links?: string[];
    published_at?: Date;
    published_by?: string;
    ordinal?: number;
    questions?: Record<string, any>;
    sub_type?: string;
    partner_organization_id?: string;
    background_image?: Record<string, any>;
    cover_image?: Record<string, any>;
    details?: Record<string, any>;
    ends_at?: Date;
    deal_ratings?: Record<string, any>;
    deal_fee?: number;
    ratings_on?: boolean;
    value?: Record<string, any>;
    patrons?: string[];
    experts?: string[];
    lead_investor_id?: string;
    partner_type?: string;
    stream_url?: string;
    related_items?: string[];
    universe?: string;
    updated_at?: Date;
    private?: boolean;
    website?: string;
    tags?: Record<string, any>;
    currency?: string;
    minimum_raise?: number;
    maximum_raise?: number;
    committed_volume?: number;
    ticket_size?: number;
    facts?: Record<string, any>;
    structure?: string[];
} 